feat: checkpoints

+++ Assets/Scripts/Source/DTOs/Request/RankingDto.cs
@@ -3,6 +3,6 @@ namespace Source.DTOs.Request
     public class RankingDto
     {
         public int score;
-        public float distance;
+        public int distance;
     }
 }
