+++ Assets/Scripts/UI/MainMenu/DTOs/SaveUserInfoDto.cs
@@ -8,7 +8,7 @@ namespace UI.DTOs
         public string username;
         public string email;
         public int score;
-        public float distance;
+        public int distance;
         // public string password;
     }
 }
